chore: Bump rust to 1.66.1

Motivation: https://blog.rust-lang.org/2023/01/10/cve-2022-46176.html

Process: Most clippy issues I fixed by running `cargo clippy --fix -- -D warnings -D clippy::all -D clippy::mem_forget -A clippy::redundant_closure -A clippy::too_many_arguments -C debug-assertions=off`. Some of these issues were not automatically fixable so I manually applied the suggestion. 

See merge request dfinity-lab/public/ic!10017

+++ rs/bitcoin/canister/src/store.rs
@@ -147,8 +147,7 @@ fn get_utxos_from_chain(
         None => all_utxos,
         // There's some limit, so use it to chunk up the UTXOs if they don't fit.
         Some(utxo_limit) => {
-            let (utxos_to_return, rest) =
-                all_utxos.split_at(all_utxos.len().min(utxo_limit as usize));
+            let (utxos_to_return, rest) = all_utxos.split_at(all_utxos.len().min(utxo_limit));
 
             if !rest.is_empty() {
                 next_page = Some(
